The replay in glvdebug can now be paused, continued, and stopped; the current packet index is made available when the replay is paused or stopped. Instead of handling a lot of the replaying code (which is API-agnostic) inside glvdebug_xgl_QController, it is now being done inside glvdebug_QReplayWorker - this includes loading the replayers and interating over the trace packets, and emitting signals.

+++ tools/glave/src/glvdebug/glvdebug_trace_file_utils.h
@@ -54,7 +54,7 @@ struct glvdebug_trace_file_info
     glv_trace_file_header header;
 
     // number of packets in file which should also be number of elements in pPacketOffsets array
-    unsigned int packetCount;
+    uint64_t packetCount;
 
     // array of packet offsets
     glvdebug_trace_file_packet_offsets* pPacketOffsets;
